Fixed player looking in random direction
Fixed player looking in random direction when the raycast plane was not at world elevation. Instead of casting a ray from a fixed plane, we now cast a ray from the camera to the world. The player than looks at the point where the ray collides with the world. This should work on all maps regardless of elevation.