Character selected in main menu now carries over between scenes
Character selected in main menu now carries over between scenes.

In the MainMenu scene, there is a new GameObject called "PlayerParent". It contains a script for spawning the player in every map. After a character is selected from the menu, the PlayerParent also gets the selected Character added. These are not destroyed on loading a new scene.